requirements.txt
mediapipe==0.10.18
nbstripout==0.8.1
numpy==1.26.4
opencv-python==4.10.0.84
pandas==pandas
pyarrow==18.1.0
pygame==2.6.1
scikit-learn==1.5.2
scikit-learn-intelex==2025.0.0
tqdm==4.67.1
ultralytics==8.3.40
python-dotenv==1.0.1